a proposal for a headless version
