+++ mydata/mydata_console.py
@@ -1,0 +1,7 @@
+class MyData(object):
+
+    def MainLoop(self):
+        """
+        some non-wx event loop
+        """
+        pass
